requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9  # PostgreSQL driver for production

# Auth
bcrypt>=4.1.0
itsdangerous>=2.1.0

# Excel processing
openpyxl>=3.1.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0

# Web scraping / search
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google Sheets
gspread>=6.1.0
google-auth>=2.29.0

# AI
anthropic>=0.18.0
google-genai>=0.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
